pyproject.toml

[project]
name = "prematch-trader"
version = "0.2.0"
description = "Sistema di trading sportivo pre-match"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pandas>=2.0.0","numpy>=1.24.0","pyyaml>=6.0"]
[project.scripts]
prematch-trader = "prematch_trader.cli:main"
[build-system]
requires = ["setuptools","wheel"]
build-backend = "setuptools.build_meta"
